feat(transaction): isolate upgrade/upload tests into separate file
